fix sql data, six delete plan

+++ SwimmingPoolServiceExample/src/main/java/com/inbadevs/swimmingpoolserviceusers/buisness/ManagerPlan.java
@@ -38,7 +38,7 @@ public class ManagerPlan {
         this.plan.update(plan);
     }
 
-    public void deletePlan(Integer id) throws BuisnessLayerException {
+    public void deletePlan(Long id) throws BuisnessLayerException {
         this.plan.delete(id);
     }
